feat: rename routes dir for modules and fix connection postgresql
