+++ src/app.module.ts
@@ -1,16 +1,17 @@
 import { ConfigModule } from "@nestjs/config";
 import { Module } from "@nestjs/common";
-import { DatabaseModule } from "./dataAccess/databases";
-import { RecluterModule } from "./routes/index.routes";
-import { MongoDatabaseModule } from "./dataAccess/databases/mongodb/database.module";
-import { StatsModule } from "./routes/stats/stats.module";
+import { 
+  RecluterModule, 
+  StatsModule
+} from "./modules/index.modules";
+import { DatabaseModule } from "./dataAccess/databases/database.module";
 @Module({
   imports: [
     ConfigModule.forRoot({ isGlobal: true }),
     RecluterModule,
     StatsModule,
     DatabaseModule,
-    MongoDatabaseModule,
+    DatabaseModule,
   ],
 })
 export class AppModule {}
