+++ src/modules/index.modules.ts
@@ -1,0 +1,2 @@
+export * from './recluter/recluter.module'
+export * from './stats/stats.module'
