+++ src/modules/recluter/interfaces/sequence.interface.ts
