--- src/routes/index.routes.ts
@@ -1,1 +1,0 @@
-export * from './recluter/recluter.module'
